Issues with seqNo and Eventhub Retention
It feels like we're having some issues with the retention of the Event Hub.
When starting a Spark job that reads from a consumer group that already has some pruned messages
because then run out of retention.

java.lang.IllegalStateException: In partition 1 of eventhubdev, with consumer group consumergroup-1,
request seqNo 1921738 is less than the received seqNo 1924294.
The earliest seqNo is 1924293 and the last seqNo is 1944993